Correcting WeatherStation class and validate Wind direction
Activate Pull up on PA8 and PA15

+++ dev/Core/Inc/WeatherStation.h
@@ -12,28 +12,28 @@
 // Definition of a simulated database object
 typedef struct T_GraphicsData {
     // To store the last 24 acquisitions
-    float temperatures[GRAPHICS_SIZE];
-    float humidities[GRAPHICS_SIZE];
-    float pressures[GRAPHICS_SIZE];
-    float rainfalls[GRAPHICS_SIZE];
-    float wind_speeds[GRAPHICS_SIZE];
-    float wind_directions[GRAPHICS_SIZE];
+    double temperatures[GRAPHICS_SIZE];
+    double humidities[GRAPHICS_SIZE];
+    double pressures[GRAPHICS_SIZE];
+    double rainfalls[GRAPHICS_SIZE];
+    double wind_speeds[GRAPHICS_SIZE];
+    double wind_directions[GRAPHICS_SIZE];
 } T_GraphicsData;
 
 T_GraphicsData GraphicsData(); // T_GraphicsData constructor
-void addGraphicsData(T_GraphicsData* gd, float temp, float hum, float press, float rain, float wspeed, float wdir); // add all acquisitions in database
+void addGraphicsData(T_GraphicsData* gd, double temp, double hum, double press, double rain, double wspeed, double wdir); // add all acquisitions in database
 
 
 // Definition of DataToSave
 typedef struct T_DataToSave {
 
     // To store the not saved data
-    float temperatures[DATA_SIZE];
-    float humidities[DATA_SIZE];
-    float pressures[DATA_SIZE];
-    float rainfalls[DATA_SIZE];
-    float wind_speeds[DATA_SIZE];
-    float wind_directions[DATA_SIZE];
+	double temperatures[DATA_SIZE];
+	double humidities[DATA_SIZE];
+	double pressures[DATA_SIZE];
+	double rainfalls[DATA_SIZE];
+	double wind_speeds[DATA_SIZE];
+	double wind_directions[DATA_SIZE];
 
     // to know how much data we have
     int nb_data;
@@ -41,22 +41,22 @@ typedef struct T_DataToSave {
 } T_DataToSave;
 
 T_DataToSave DataToSave(); // T_DataToSave constructor
-void addDataToSave(T_DataToSave * ds, float temp, float hum, float press, float rain, float wspeed, float wdir); // add all data to save
+void addDataToSave(T_DataToSave * ds, double temp, double hum, double press, double rain, double wspeed, double wdir); // add all data to save
 void clearBuffers(T_DataToSave* ds);
 
 // Definition of the Weather Station object
 typedef struct T_WeatherStation {
-    float temperature;
-    float humidity;
-    float pressure;
-    float rainfall;
-    float wind_speed;
-    float wind_direction; // degrees
+	double temperature;
+	double humidity;
+	double pressure;
+	double rainfall;
+	double wind_speed;
+	double wind_direction; // degrees
 } T_WeatherStation;
 
 T_WeatherStation WeatherStationDefault(); // Default Constructor
-T_WeatherStation WeatherStation(float temp, float hum, float press, float rain, float wspeed, float wdir); // Constructor with args
-void updateWeatherStation(T_WeatherStation* ws, T_GraphicsData* gd, T_DataToSave * ds, float temp, float hum, float press, float rain, float wspeed, float wdir); // Update the current values of the WeatherStation
+T_WeatherStation WeatherStation(double temp, double hum, double press, double rain, double wspeed, double wdir); // Constructor with args
+void updateWeatherStation(T_WeatherStation* ws, T_GraphicsData* gd, T_DataToSave * ds, double temp, double hum, double press, double rain, double wspeed, double wdir); // Update the current values of the WeatherStation
 
 
 // Constants
